fix: explicitly specify dtype for discrete_action and history_mask tensors
- Add torch.long for discrete_action in tokenized_dataset.py and packing_collator.py
- Add torch.bool for history_mask in packing_collator.py

+++ src/data/tokenized_dataset.py
@@ -367,7 +367,7 @@ class FlattenedCollatorForVLN:
             "heatmap": torch.stack([inst['heatmap'] for inst in instances], dim=0),
             "action": torch.stack([inst['action'] for inst in instances], dim=0),
             "action_valid": torch.tensor([inst['action_valid'] for inst in instances]),
-            "discrete_action": torch.tensor([inst.get('discrete_action', 1) for inst in instances]),
+            "discrete_action": torch.tensor([inst.get('discrete_action', 1) for inst in instances], dtype=torch.long),
             "is_stop": torch.tensor([inst.get('is_stop', 0.0) for inst in instances]),
             "is_flipped": torch.tensor([inst.get('is_flipped', False) for inst in instances], dtype=torch.bool),
             "text": [inst['text'] for inst in instances],
